Add Django built-in CSP nonce support (#2267)
Django 6.0 added built-in support for Content Security Policies.
Part of this support is associating a randomly generated nonce to each request that can then be
attached to <script> and <style> tags.

django-debug-toolbar already has support for CSP nonces generated by the django-csp third-party lib but
the nonces generated by django-csp and Django are accessed differently.

This commit adds support for the CSP nonces generated by the built-in Django implementation.