Update the hash and version of python-coverage-comment-action (#2269)
This specifies the minor version number since the action seems to update
the hash for the major version tag. This seems to be throwing Zizmor off
and is causing warnings. This means we may not have the latest commits, but
the maintainers are doing regular releases, so that should be fine.